change some uint64_t to uint32_t in ClientSocket

+++ ClientSocket.cpp
@@ -26,7 +26,7 @@
 
 #include <stdio.h>
 
-ClientSocket::ClientSocket(uint64_t addr, uint16_t port, int32_t style)
+ClientSocket::ClientSocket(uint32_t addr, uint16_t port, int32_t style)
 	: mSock(socket(AF_INET, style, 0))
 {
 	memset(&mSockAddr, 0, sizeof(mSockAddr));
@@ -40,7 +40,7 @@ ClientSocket::~ClientSocket()
 	close(mSock);
 }
 
-TcpSocket::TcpSocket(uint64_t addr, uint16_t port)
+TcpSocket::TcpSocket(uint32_t addr, uint16_t port)
 	: ClientSocket(addr, port, SOCK_STREAM)
 {
 	if(connect(mSock, reinterpret_cast<sockaddr*>(&mSockAddr), sizeof(mSockAddr)) < 0)
@@ -87,7 +87,7 @@ int32_t TcpSocket::Send(const uint8_t* frame, size_t length) const
 	return send(mSock, frame, length, 0);
 }
 
-UdpSocket::UdpSocket(uint64_t addr, uint16_t port, bool doBind, int32_t enableBroadcast)
+UdpSocket::UdpSocket(uint32_t addr, uint16_t port, bool doBind, int32_t enableBroadcast)
 	: ClientSocket(addr, port, SOCK_DGRAM)
 {
 	setsockopt(mSock, SOL_SOCKET, SO_BROADCAST, &enableBroadcast, sizeof(enableBroadcast));
